feat: LobbySettings rendering test for different lobbies refactor: login helper to reduce repetitive code fix: knowledge mode community toggle
