+++ playwright.config.js
@@ -3,6 +3,7 @@ import { defineConfig, devices } from "@playwright/test";
 export default defineConfig({
   testDir: "./tests",
   timeout: 30_000,
+  workers: 1,
   use: {
     baseURL: "http://localhost:5173",
     headless: true,
